Update frontend to use localstorage for key instead of cookies
