+++ packages/app/src/providers/localStorageAuthProvider.tsx
@@ -1,0 +1,20 @@
+import { useGetLocalApiKey } from '../api/localAuth.api'
+import { LoadingScreen } from '../components/LoadingScreen'
+import authUtil from '../utils/auth.util'
+
+interface Props {
+  children: React.ReactNode
+}
+
+export const LocalStorageAuthProvider = ({ children }: Props) => {
+  const localApiKey = authUtil.getLocalApiKey()
+  const { data, isFetching } = useGetLocalApiKey(!localApiKey)
+
+  if (isFetching) {
+    return <LoadingScreen />
+  }
+  if (data) {
+    authUtil.setLocalApiKey(data.apiKey)
+  }
+  return <>{children}</>
+}
