+++ packages/app/src/routes/AppRoutes.tsx
@@ -1,12 +1,11 @@
 import { Route, Routes } from 'react-router-dom'
 
-import PublicLayout from '../layouts/PublicLayout'
 import { HomePage } from '../components/Home/HomePage'
 import { useAnalyticsPageSubscription } from '../hooks/useAnalytics'
 import InstallPage from '../components/InstallPage'
 import ImportLayout from '../layouts/ImportLayout'
 import { ImportPage } from '../components/ImportPage'
-import BaseLayout from '../layouts/BaseLayout'
+import DashboardLayout from '../layouts/DashboardLayout'
 
 // We want to subscribe to events in the router, but we don't want to rerender
 // the entire app when the route changes.
@@ -21,17 +20,13 @@ export const AppRoutes = () => {
     <>
       <AppRoutesPageSubscription />
       <Routes>
-        <Route element={<BaseLayout />}>
-          <Route element={<PublicLayout />}>
-            <Route path="/" element={<HomePage />} />
-          </Route>
-          <Route element={<ImportLayout />}>
-            <Route path="/import" element={<ImportPage />} />
-          </Route>
-          <Route>
-            <Route path="/install" element={<InstallPage />} />
-          </Route>
+        <Route element={<DashboardLayout />}>
+          <Route path="/" element={<HomePage />} />
         </Route>
+        <Route element={<ImportLayout />}>
+          <Route path="/import" element={<ImportPage />} />
+        </Route>
+        <Route path="/install" element={<InstallPage />} />
       </Routes>
     </>
   )
